full_requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
autopep8==2.0.2
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
breadability==0.1.20
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
ckip-transformers==0.3.4
click==8.1.3
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1670575068857/work
confection==0.0.4
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1674522353161/work
decorator==5.1.1
dill==0.3.6
docopt==0.6.2
entrypoints==0.4
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastjsonschema @ file:///home/conda/feedstock_root/build_artifacts/python-fastjsonschema_1677336799617/work/dist
filelock==3.9.1
fonttools==4.38.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1676919000169/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1676322140253/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1677617093347/work
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1671720089366/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1676579893731/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1675109852367/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1671722028097/work
kiwisolver==1.4.4
langchain==0.0.190
langcodes==3.3.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1673560067442/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.27.7
openapi-schema-pydantic==1.2.4
p-tqdm==1.4.0
packaging==23.1
pandas==2.0.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathos==0.3.0
pathy==0.10.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1633981968097/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1677883256757/work
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.8
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==4.23.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1667885909637/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==12.0.0
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1673612650067/work
regex==2023.6.3
requests==2.28.2
rich==13.4.1
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-pkuseg==0.0.32
SQLAlchemy==2.0.15
srsly==2.4.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.23.1
streamlit-chat==0.0.2.2
sumy==0.11.0
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
tornado==6.3.2
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.26.1
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
weaviate-client==3.19.2
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1672066693230/work
yarg==0.1.9
yarl==1.9.2
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.5.0/zh_core_web_sm-3.5.0-py3-none-any.whl
zipp==3.15.0
